pyproject.toml
[project]
name = "cooking-agent"
version = "0.1.0"
description = "Self-improving cooking agent — LangWatch × Flagsmith"
requires-python = ">=3.11"
dependencies = [
    "agno>=0.2.0",
    "openai>=1.50.0",
    "langwatch>=0.2.0",
    "langwatch-scenario>=0.7.0",
    "flagsmith>=3.8.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]

[project.scripts]
cooking-agent = "agent.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["agent_test: scenario-based agent tests"]
